requirements.txt
alembic==1.11.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
defusedxml==0.7.1
elastic-transport==8.4.0
elasticsearch==8.9.0
exceptiongroup==1.1.3
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
httpie==3.2.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
Pygments==2.15.1
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.2
pytest-mock==3.11.1
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
SQLAlchemy==2.0.19
tomli==2.0.1
types-Flask-Migrate==4.0.0.6
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
